User details page is pre-populated with info
Also configures API to use Access Token and Backend to use ID Token.
This is needed because we use the Organization Server in Okta, which only returns group claims in ID tokens. We use this server because it allows issuing tokens with okta.users.manage.self scope, which allows users to edit their own details/password.