Sets up different properties based on Azure vs. local profile
Now we can swap between the Azure properties and the local by changing the active profile.

This warranted adding the corsConfigurationSource Bean in SecurityConfig. It sets up CORS with the local Angular frontend. and is only created when run in the "local" profile. When run in "prod," the app is in Azure, and Azure configures Cross-Origin itself.

Before, we had to configure @CrossOrigin annotations in each controller, but this bean provides a global CORS configuration.