Fixes backend 500 StackOverflow error
I believe it was caused by us directly saving the product sent in a cart item.

This product had a null list of ratings... I'm not 100% clear why, but saving a null List<ProductRating> seems to mess with the ToString.exclude (and tested EqualsAndHashCode.Exclude) annotations.
In other cases these annotations prevent StackOverflow errors caused by an OrderItem referencing a Order, that references a OrderItem -> Order -> OrderItem infinite times.

Now, we get the product id from the cart item, and find it in the repo.
Then we update the product we find from the repo. The product in the repo holds the ratings, whereas the cart item does not.